Move FederationSupportTest and PendingFederationTest from federation package to match main packages
